+++ src/components/Button/Button.story.js
@@ -52,13 +52,14 @@ stories
     return (
       <ButtonRow
         centerGroup={ [
-          <Button key="1" isActive={ boolean('isActive', false) } width="default">Button</Button>,
-          <Button key="2" isActive={ boolean('isActive', false) } style="highlight" width="default">Highlight</Button>,
-          <Button key="3" isActive={ boolean('isActive', false) } style="danger" width="default">Danger</Button>,
-          <Button key="4" isActive={ boolean('isActive', false) } style="danger-outline" width="default">Danger Outline</Button>,
-          <Button key="5" isActive={ boolean('isActive', false) } style="outline" width="default">Outline</Button>,
-          <Button key="6" isActive={ boolean('isActive', false) } style="plain" width="default">Plain</Button>,
-          <Button key="7" isActive={ boolean('isActive', false) } style="toggle" width="default">Toggle</Button>,
+          <Button key="1" isActive={ false } width="default">Button</Button>,
+          <Button key="2" isActive={ false } style="highlight" width="default">Highlight</Button>,
+          <Button key="3" isActive={ false } style="danger" width="default">Danger</Button>,
+          <Button key="4" isActive={ false } style="danger-outline" width="default">Danger Outline</Button>,
+          <Button key="5" isActive={ false } style="outline" width="default">Outline</Button>,
+          <Button key="6" isActive={ false } style="plain" width="default">Plain</Button>,
+          <Button key="7" isActive={ boolean('Toggle button isActive', false) } style="toggle" width="default">Toggle</Button>,
+          <Button key="8" isActive={ false } style="underline" width="default">Underline</Button>,
         ] }
       />);
   }))
@@ -73,10 +74,11 @@ stories
           <Button key="5" isDisabled={ true } style="outline" width="default">Outline</Button>,
           <Button key="6" isDisabled={ true } style="plain" width="default">Plain</Button>,
           <Button key="7" isDisabled={ true } style="toggle" width="default">Toggle</Button>,
+          <Button key="8" isDisabled={ true } style="underline" width="default">Underline</Button>,
         ] }
       />);
   })
-  .add('Underline button', () => {
+  .add('Underline button with Arrow Icon', () => {
     return (
       <Button style="underline" width="default">
         Underline Button
@@ -114,7 +116,7 @@ stories
       </Button>
     );
   }))
-  .add('Button that disables on click', () => {
+  .add('Loading Button that disables on click', () => {
     return (
       <Button
         style="highlight"
@@ -124,7 +126,7 @@ stories
       </Button>
     );
   })
-  .add('Button that disables on click with custom text', () => {
+  .add('Loading Button that disables on click with custom text', () => {
     return (
       <Button
         style="highlight"